Duration (#2)
* add recipe timings

and related schema microformats

* Add basic gemfile / .gitignore

* Put timings in component recipes

* handle lowercase designators

* Add prepTime and cookTime supprt

Wholly via javascript with no dependencies. This allows for it to be fully compatible with gh_pages

* Add prepTime and cookTime supprt

Wholly via javascript with no dependencies. This allows for it to be fully compatible with gh_pages

* Tolerate lowercase

Allow users to use lowercase in the definitions. This can lead to better readability:

preptime:  1h20m

* cleanup

* unique dom id for timings

allows for multiple timings per page like when using components

* include schema totalTime

Co-authored-by: Rue Turner <rue@thinlayer.co.uk>